compile on window

+++ src/c_lib/t_map/t_vbo_draw.cpp
@@ -631,7 +631,7 @@ void draw_vbo_debug(int x, int y)
         if(i==cx && j == cy) 
             glColor3ub(255, 255, 0);
 
-        glVertex3f(x+(psize+sep)*i+off, yoff_1+y+(psize+sep)*j+off, -0.1);
+        glVertex3f( (float) x+(psize+sep)*i+off, (float) yoff_1+y+(psize+sep)*j+off, -0.1f);
     }
     glEnd();
